feat(common): reorganize Verbs and formalize arguments
All verbs are now moved into their own package to make the model
classes more comprehensable. It is more clear now, what these are for.

The Verb interface has been extended:
1. To carry an enumerated list of OAI-PMH specified Arguments (copied and refactored from
   the xoai-data-provider/OAIRequest.Parameters enum)
2. Associate each verb type within Verb.Type enum with sets of required,
   optional and exclusive arguments, following the OAI-PMH 2.0 spec.

This will allow for better validation when parsing queries if all
arguments are present and allowed. (It does not yet add validation of
the argument value)