feat!(common): make Request model more specific, extend XmlWriter
Instead of using different instance values for the attributes of <request>,
use an EnumMap to store them, reusing the Verb.Argument enum introduced earlier.
Instead of returning null values, use the Optional util to make the API more fluent.

Reuse this to extend XmlWriter with a writeAttribute() method taking Verb.Arguments and Optional
as arguments - unwrapping happens in there, lowering the complexity of the Request.write().
This also avoid repeated strings for the attribute name, as we reuse already checked and defined ones
from the Verb.Argument enum.