feat(common): make ResumptionTokenFormat granularity configurable
When encoding or decoding the token, we need to pass a granularity of the timestamps
that are acceptable. Now the format interface requires a method to configure
the granularity and return the format instance, leaving it open to a using code
to go with the default of the format or provide something.

Tests have been added for SimpleResumptionTokenFormat.