feat(common): make XmlWriter.WriterContext an interface
To be able to override the XmlWriter settings for granularity and
resumption token encoding/decoding, the existing inner class has been
moved to an interface with default methods.

This makes the default still second granularity and use the SimpleResumptionTokenFormat,
but will allow the xoai-dataprovider/RepositoryConfiguration to serve as a WriterContext,
too. This way, when creating the data provider, the necessary creation of the
XmlWriter will allow reusing the also existing RepositoryConfiguration and avoid double
efforts.