feat(common): extract token via ResumptionTokenFormat.parse(Request)
The interface now features a default implementation of how to parse
a ResumptionToken.Value using the format and the parse(String) method.
As the request might carry no token, it will return an empty Optional
if no token present. It will throw an exception when the token
cannot be parsed.