refactor(data): migrate and simplify ListSetsHandler and SetRepository
With this commit, the application provided SetRepository interface is relaxed
to provide a simple returning point of all sets present in the repository.
The library itself provides the necessary logic to create paginated results
of sets to drive down the complexity of the ListSetsHandler.

It is expected that the number of sets is low and dynamic results may be cached
by the implementing repository. The overhead of creating the paginated result ourselves
is negligible. There is not much benefit in storing paginated results within the
set repository, as the model is rather small (compared to the number of records
and their metadata)