feat(data): make <metadata> attributes configurable #32
Necessary for Dataverse 4/5 compatibility. They use a special
metadata format "dataverse_json" which exposes the records
full JSON API query URL as a <metadata> attribute.

This is non-OAI-PMH spec conformant!
As the RepositoryConfiguration can be used as a WriterContext,
this enables an application (Dataverse) to make an XmlWriter
(which the application creates anyway) write out the attributes.

It's up to the application how to enable this in the configuration.
Be it static in code or as a configuration option.

This commit also adds a test to ensure this is working for
get/list record verbs.