fix(common): apply more lowlevel fix for broken UTF-8 chars #188
With this commit, we introduce an analysis routine that will go over
the last few bytes when reading with CopyElement. This is faster
than converting to String and checking for the UTF-8 unknown char sign
over and over again.

By using a buffered input stream, we can rewind the stream if necessary
and read again up to the point that we don't have a broken char.

Extensive testing was added to make sure the analysis function works
with any length of multibyte UTF-8 chars